Added the PiecewisePolynomial class
Added the PiecewisePolynomial class that represents a PiecewisePolynomial function

+++ Math/PiecewisePolynomialPart.hpp
@@ -50,7 +50,7 @@ namespace LouLib {
 
             double getEnd() const;
 
-            double evaluate(int t);
+            double evaluate(double t);
         };
 
     } // LouLib
